Gyro upplösning
Högre upplösning på gyrot. 10 bitar istället för 8 på adc. Vet ej om det
funkar, men vi får hoppas! Bara att testa på!

Även lagt till så att gyro_null får nytt värde ifall man kalibrerar
skiten. Dvs, gyro-värdet när ekpitaget inte svänger nåt. Helst ska den
stå helt still.

+++ Kod/mainsensor_real/mainsensor/mainsensor/gyro.h
@@ -1,6 +1,6 @@
 #ifndef GYRO_H
 #define GYRO_H
 
-void read_gyro(const uint8_t wanted_degrees, const uint8_t gyro_null);
+void read_gyro(const uint8_t wanted_degrees, const uint16_t gyro_null);
 
 #endif
